refactor: Saga 엔티티 및 enum을 infrastructure 레이어로 이동
