+++ order-service/src/main/java/com/palja/order_service/application/saga/steps/CreatePaymentStep.java
@@ -6,7 +6,7 @@ import com.palja.order_service.application.exception.OrderErrorCode;
 import com.palja.order_service.application.port.PaymentClient;
 import com.palja.order_service.application.saga.SagaStep;
 import com.palja.order_service.domain.entity.Order;
-import com.palja.order_service.domain.vo.OrderSagaStep;
+import com.palja.order_service.infrastructure.saga.model.OrderSagaStep;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
