+++ order-service/src/main/java/com/palja/order_service/application/service/impl/OrderSagaServiceImpl.java
@@ -3,7 +3,7 @@ package com.palja.order_service.application.service.impl;
 import com.palja.common.exception.BusinessException;
 import com.palja.order_service.application.exception.OrderErrorCode;
 import com.palja.order_service.application.service.OrderSagaService;
-import com.palja.order_service.domain.entity.OrderSaga;
+import com.palja.order_service.infrastructure.saga.model.OrderSaga;
 import com.palja.order_service.domain.repository.OrderSagaRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
