+++ order-service/src/main/java/com/palja/order_service/infrastructure/saga/model/OrderSagaStep.java
@@ -1,4 +1,4 @@
-package com.palja.order_service.domain.vo;
+package com.palja.order_service.infrastructure.saga.model;
 
 import lombok.Getter;
